Fix: customFetch 세션 만료 처리 테스트 코드를 위한 process.env.TEST_ENV !== SSR 조건 추가

+++ shared/common/utils/customFetch.ts
@@ -27,7 +27,8 @@ export const createFetch = (baseURL: string, defaultOptions?: RequestInit) => {
     if (
       response.status === 401 &&
       data?.message === "만료된 세션이에요." &&
-      typeof window !== "undefined"
+      typeof window !== "undefined" &&
+      process.env.TEST_ENV !== "SSR"
     ) {
       await fetch("/api/auth/session-cookie", { method: "DELETE" });
       redirect("/session-expired");
